Update README.md
This repository is a collection of my hands-on work with MySQL, covering everything from basic SQL queries to more advanced concepts like joins, subqueries, data modeling, and structured reporting.

I created this project as part of my learning journey in databases and data analysis. It’s not just a single script, but rather a set of realistic datasets and SQL queries that reflect how data problems are solved in real business scenarios such as retail, e-commerce, and supply chain management.